+++ common/modules/rewards/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "rewards"
+version = "0.0.0"
+authors = [ "catalin-neagu <catalin.neagu@elrond.com>",]
+edition = "2018"
+
+[lib]
+path = "src/rewards.rs"
+
+[dependencies.common_structs]
+path = "../../common_structs"
+
+[dependencies.config]
+path = "../config"
+
+[dependencies.farm_token]
+path = "../farm_token"
+
+[dependencies.token_send]
+path = "../token_send"
+
+[dependencies.elrond-wasm]
+version = "0.24.0"
