+++ common/modules/rewards/src/rewards.rs
@@ -1,3 +1,5 @@
+#![no_std]
+
 elrond_wasm::imports!();
 elrond_wasm::derive_imports!();
 
@@ -5,7 +7,7 @@ use common_structs::Nonce;
 
 #[elrond_wasm::module]
 pub trait RewardsModule:
-    config::ConfigModule + token_send::TokenSendModule + farm_token::FarmTokenModule
+    config::ConfigModule + farm_token::FarmTokenModule + token_send::TokenSendModule
 {
     fn calculate_per_block_rewards(
         &self,
@@ -20,31 +22,6 @@ pub trait RewardsModule:
         let block_nonce_diff = current_block_nonce - last_reward_block_nonce;
 
         per_block_reward * block_nonce_diff
-    }
-
-    fn mint_per_block_rewards(&self, token_id: &TokenIdentifier) -> BigUint {
-        let current_block_nonce = self.blockchain().get_block_nonce();
-        let last_reward_nonce = self.last_reward_block_nonce().get();
-
-        if current_block_nonce > last_reward_nonce {
-            let to_mint = self.calculate_per_block_rewards(current_block_nonce, last_reward_nonce);
-
-            if to_mint != 0 {
-                self.send().esdt_local_mint(token_id, 0, &to_mint);
-            }
-            self.last_reward_block_nonce().set(&current_block_nonce);
-            to_mint
-        } else {
-            BigUint::zero()
-        }
-    }
-
-    fn generate_aggregated_rewards(&self, reward_token_id: &TokenIdentifier) {
-        let total_reward = self.mint_per_block_rewards(reward_token_id);
-        if total_reward > 0 {
-            self.increase_reward_reserve(&total_reward);
-            self.update_reward_per_share(&total_reward);
-        }
     }
 
     fn increase_reward_reserve(&self, amount: &BigUint) {
@@ -106,25 +83,6 @@ pub trait RewardsModule:
         let current_nonce = self.blockchain().get_block_nonce();
         self.produce_rewards_enabled().set(&true);
         self.last_reward_block_nonce().set(&current_nonce);
-        Ok(())
-    }
-
-    #[endpoint]
-    fn end_produce_rewards(&self) -> SCResult<()> {
-        self.require_permissions()?;
-        let reward_token_id = self.reward_token_id().get();
-        self.generate_aggregated_rewards(&reward_token_id);
-        self.produce_rewards_enabled().set(&false);
-        Ok(())
-    }
-
-    #[endpoint(setPerBlockRewardAmount)]
-    fn set_per_block_rewards(&self, per_block_amount: BigUint) -> SCResult<()> {
-        self.require_permissions()?;
-        require!(per_block_amount != 0, "Amount cannot be zero");
-        let reward_token_id = self.reward_token_id().get();
-        self.generate_aggregated_rewards(&reward_token_id);
-        self.per_block_reward_amount().set(&per_block_amount);
         Ok(())
     }
