+++ dex/farm_with_lock/src/lib.rs
@@ -3,8 +3,8 @@
 #![feature(exact_size_is_empty)]
 
 pub mod custom_config;
+mod custom_rewards;
 pub mod farm_token_merge;
-mod rewards;
 
 use common_structs::{Epoch, FarmTokenAttributes, Nonce};
 use config::State;
@@ -27,7 +27,8 @@ type ExitFarmResultType<BigUint> =
 
 #[elrond_wasm::contract]
 pub trait Farm:
-    rewards::RewardsModule
+    custom_rewards::CustomRewardsModule
+    + rewards::RewardsModule
     + custom_config::CustomConfigModule
     + config::ConfigModule
     + token_send::TokenSendModule
@@ -115,7 +116,7 @@ pub trait Farm:
 
         let farm_contribution = &enter_amount;
         let reward_token_id = self.reward_token_id().get();
-        self.generate_aggregated_rewards(&reward_token_id);
+        self.generate_aggregated_rewards();
 
         let epoch = self.blockchain().get_block_epoch();
         let attributes = FarmTokenAttributes {
@@ -177,7 +178,7 @@ pub trait Farm:
 
         let farm_attributes = self.get_farm_attributes(&payment_token_id, token_nonce)?;
         let mut reward_token_id = self.reward_token_id().get();
-        self.generate_aggregated_rewards(&reward_token_id);
+        self.generate_aggregated_rewards();
 
         let mut reward = self.calculate_reward(
             &amount,
@@ -270,7 +271,7 @@ pub trait Farm:
         let farm_attributes = self.get_farm_attributes(&payment_token_id, token_nonce)?;
 
         let mut reward_token_id = self.reward_token_id().get();
-        self.generate_aggregated_rewards(&reward_token_id);
+        self.generate_aggregated_rewards();
 
         let mut reward = self.calculate_reward(
             &amount,
@@ -378,7 +379,7 @@ pub trait Farm:
             farming_token == reward_token,
             "Farming token differ from reward token"
         );
-        self.generate_aggregated_rewards(&reward_token);
+        self.generate_aggregated_rewards();
 
         let current_rps = self.reward_per_share().get();
         let farm_attributes = self.get_farm_attributes(&payment_token_id, payment_token_nonce)?;
